+++ src/Yamua/ExecuteAndForget.cs
@@ -15,25 +15,25 @@ namespace Yamua
         /// <summary>
         /// Execute a file (exe, link etc.). Windows will display an error message when something goes wrong.
         /// </summary>
-        /// <param name="FilenamePath">Filename and path to the file to be executed, a WWW link will work also</param>
+        /// <param name="fileNamePath">Filename and path to the file to be executed, a WWW link will work also</param>
         /// <returns>TRUE if the file was started successfully</returns>
-        public static bool Execute(string FilenamePath)
+        public static bool Execute(string fileNamePath)
         {
-            return Execute(FilenamePath, string.Empty);
+            return Execute(fileNamePath, string.Empty);
         }
 
         /// <summary>
         /// Execute a file (exe, link etc.). Windows will display an error message when something goes wrong.
         /// </summary>
-        /// <param name="FilenamePath">Filename and path to the file to be executed, a WWW link will work also</param>
-        /// <param name="Arguments">Arguments to be used</param>
+        /// <param name="fileNamePath">Filename and path to the file to be executed, a WWW link will work also</param>
+        /// <param name="arguments">Arguments to be used</param>
         /// <returns>TRUE if the file was started successfully</returns>
-        public static bool Execute(string FilenamePath, string Arguments)
+        public static bool Execute(string fileNamePath, string arguments)
         {
             ProcessStartInfo psi = new ProcessStartInfo();
             psi.ErrorDialog = true;
-            psi.FileName = FilenamePath;
-            psi.Arguments = Arguments;
+            psi.FileName = fileNamePath;
+            psi.Arguments = arguments;
 
             try
             {