+++ src/Yamua/PathExtension.cs
@@ -16,16 +16,16 @@ namespace Yamua
         /// <summary>
         /// Checks if a given string is valid, this means it does not include any chars that can not be used for a filename 
         /// </summary>
-        /// <param name="Filename">A filename to be checked</param>
+        /// <param name="fileName">A filename to be checked</param>
         /// <returns>TRUE if the Filename should be OK, FALSE otherwise</returns>
-        public static bool IsFilenameValid(string Filename)
+        public static bool IsFilenameValid(string fileName)
         {
             bool returnvalue = false;
 
             //First check if this maybe a full path (C:\Testing123\MyFile.ext)
             try
             {
-                string filename_only = Path.GetFileName(Filename);
+                string filename_only = Path.GetFileName(fileName);
                 
                 //Original code from http://stackoverflow.com/a/4650495/612954 by [Phil Hunt](http://stackoverflow.com/users/492543/phil-hunt)
                 returnvalue = !string.IsNullOrWhiteSpace(filename_only) && filename_only.IndexOfAny(Path.GetInvalidFileNameChars()) < 0;
@@ -45,50 +45,50 @@ namespace Yamua
         /// </summary>
         /// <param name="Path"></param>
         /// <returns></returns>
-        public static string FullPath(string FileOrDirectoryPath)
+        public static string FullPath(string fileOrDirectoryPath)
         {
-            return Path.GetFullPath(FileOrDirectoryPath);
+            return Path.GetFullPath(fileOrDirectoryPath);
         }
 
         /// <summary>
         /// Identifical to Path.GetFileName()
         /// </summary>
-        /// <param name="Filepath">Full path to a file</param>
+        /// <param name="filePath">Full path to a file</param>
         /// <returns>The filename, including extension, from the given path</returns>
-        public static string Filename(string Filepath)
+        public static string Filename(string filePath)
         {
-            return Path.GetFileName(Filepath);
+            return Path.GetFileName(filePath);
         }
 
         /// <summary>
         /// Identical to Path.Combine(string1, string2)
         /// </summary>
-        /// <param name="DirectoryPath"></param>
-        /// <param name="Filename"></param>
+        /// <param name="directoryPath"></param>
+        /// <param name="fileName"></param>
         /// <returns></returns>
-        public static string Combine(string DirectoryPath, string Filename)
+        public static string Combine(string directoryPath, string fileName)
         {
-            return Path.Combine(DirectoryPath, Filename);
+            return Path.Combine(directoryPath, fileName);
         }
 
         /// <summary>
         /// Identical to Directory.Exists(path)
         /// </summary>
-        /// <param name="DirectoryPath"></param>
+        /// <param name="directoryPath"></param>
         /// <returns></returns>
-        public static bool DirectoryExists(string DirectoryPath)
+        public static bool DirectoryExists(string directoryPath)
         {
-            return Directory.Exists(DirectoryPath);
+            return Directory.Exists(directoryPath);
         }
 
         /// <summary>
         /// Identical to File.Exists(path)
         /// </summary>
-        /// <param name="Filepath"></param>
+        /// <param name="filePath"></param>
         /// <returns></returns>
-        public static bool FileExists(string Filepath)
+        public static bool FileExists(string filePath)
         {
-            return File.Exists(Filepath);
+            return File.Exists(filePath);
         }
     }
 }