+++ src/Yamua/StringFilter.cs
@@ -11,23 +11,23 @@ namespace Yamua
         public const string CHARS_NUMERIC = "0123456789";
         public const string CHAR_MINUS = "-";
         public const string CHAR_UNDERSCORE = "_";
-        
+
 
 
         /// <summary>
         /// This function will filter (remove) all chars that are NOT INCLUDED in the given AllowedChars parameter.
         /// </summary>
-        /// <param name="Input"></param>
-        /// <param name="AllowedChars"></param>
+        /// <param name="input"></param>
+        /// <param name="allowedChars"></param>
         /// <returns></returns>
-        public static string Filter(string Input, string AllowedChars)
+        public static string Filter(string input, string allowedChars)
         {
             int iPos = -1;
             StringBuilder sbResult = new StringBuilder();
 
-            foreach (char ch in Input.ToCharArray())
+            foreach (char ch in input.ToCharArray())
             {
-                iPos = AllowedChars.IndexOf(ch);
+                iPos = allowedChars.IndexOf(ch);
                 if (iPos > -1)
                 {
                     sbResult.Append(ch);
@@ -39,23 +39,23 @@ namespace Yamua
         /// <summary>
         /// This function will ONLY remove the chars in the RemoveChars string but leave all other in place
         /// </summary>
-        /// <param name="Input"></param>
-        /// <param name="RemoveChars"></param>
+        /// <param name="input"></param>
+        /// <param name="removeChars"></param>
         /// <returns></returns>
-        public static string Remove(string Input, string RemoveChars)
+        public static string Remove(string input, string removeChars)
         {
             int iPos = -1;
             StringBuilder sbResult = new StringBuilder();
 
-            foreach (char ch in Input.ToCharArray())
+            foreach (char ch in input.ToCharArray())
             {
-                iPos = RemoveChars.IndexOf(ch);
+                iPos = removeChars.IndexOf(ch);
                 if (iPos == -1)
                 {
                     sbResult.Append(ch);
                 }
             }
-            
+
             return sbResult.ToString();
         }